Renames files package to roles for domain clarity
Moves internal/files to internal/roles and renames files.go to files.go
within the roles package. This better reflects the domain-specific purpose
of role management rather than generic file operations.
