+++ internal/roles/files.go
@@ -1,4 +1,4 @@
-package files
+package roles
 
 import (
 	"errors"
