+++ internal/roles/files_test.go
@@ -1,4 +1,4 @@
-package files
+package roles
 
 import (
 	"os"
